feat: Implement built-in OAuth templates with ejection capability
- Add oauth_templates.go with embedded default templates as constants
- Implement template override system: external templates take precedence over built-in
- Add --eject-templates CLI flag to export built-in templates for customization
- Update template loading logic with fallback mechanism
- Remove go:embed dependency - templates now work without external files
- Update documentation with new template system and ejection workflow
- Templates work out-of-the-box but can be customized when needed

Benefits:
- Zero external dependencies - OAuth works immediately after build
- Easy customization via template ejection
- Graceful fallback from external to built-in templates
- Clear upgrade path for template customization
