+++ main.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"fmt"
 	"log"
+	"os"
+	"path/filepath"
 )
 
 var BuildVersion = "dev"
@@ -13,6 +15,7 @@ func main() {
 	insecure := flag.Bool("insecure", false, "allow insecure HTTPS connections by skipping TLS certificate verification")
 	version := flag.Bool("version", false, "print version and exit")
 	help := flag.Bool("help", false, "print help and exit")
+	ejectTemplates := flag.Bool("eject-templates", false, "eject OAuth templates to templates/oauth/ directory for customization")
 	flag.Parse()
 	if *help {
 		flag.Usage()
@@ -20,6 +23,12 @@ func main() {
 	}
 	if *version {
 		fmt.Println(BuildVersion)
+		return
+	}
+	if *ejectTemplates {
+		if err := ejectOAuthTemplates(); err != nil {
+			log.Fatalf("Failed to eject templates: %v", err)
+		}
 		return
 	}
 	config, err := load(*conf, *insecure)
@@ -30,4 +39,36 @@ func main() {
 	if err != nil {
 		log.Fatalf("Failed to start server: %v", err)
 	}
+}
+
+func ejectOAuthTemplates() error {
+	templatesDir := "templates/oauth"
+	
+	// Create templates directory
+	if err := os.MkdirAll(templatesDir, 0755); err != nil {
+		return fmt.Errorf("failed to create templates directory: %v", err)
+	}
+	
+	// Write authorize.html
+	authorizePath := filepath.Join(templatesDir, "authorize.html")
+	if err := os.WriteFile(authorizePath, []byte(defaultAuthorizePage), 0644); err != nil {
+		return fmt.Errorf("failed to write authorize.html: %v", err)
+	}
+	
+	// Write success.html
+	successPath := filepath.Join(templatesDir, "success.html")
+	if err := os.WriteFile(successPath, []byte(defaultSuccessPage), 0644); err != nil {
+		return fmt.Errorf("failed to write success.html: %v", err)
+	}
+	
+	fmt.Printf("OAuth templates ejected to %s/\n", templatesDir)
+	fmt.Println("You can now customize the HTML templates and restart the server to use them.")
+	fmt.Println()
+	fmt.Println("Template files created:")
+	fmt.Printf("  %s - OAuth authorization/login page\n", authorizePath)
+	fmt.Printf("  %s - OAuth success/redirect page\n", successPath)
+	fmt.Println()
+	fmt.Println("To use the built-in templates again, simply remove the templates/ directory.")
+	
+	return nil
 }
